🐛 config only works before Part import
